900: Add field nuernberg pass id
